Refactor header logo and icons location
